Properties must be ints

+++ clientfactory-jersey/src/main/java/com/opentable/jaxrs/JaxRsClientFactoryImpl.java
@@ -36,8 +36,8 @@ public class JaxRsClientFactoryImpl implements InternalClientFactory
         final ClientConfig clientConfig = new ClientConfig();
         clientConfig.property(ApacheClientProperties.CONNECTION_MANAGER, connectionManager);
         clientConfig.connectorProvider(new ApacheConnectorProvider());
-        clientConfig.property(ClientProperties.CONNECT_TIMEOUT, config.connectTimeout().getMillis());
-        clientConfig.property(ClientProperties.READ_TIMEOUT, config.socketTimeout().getMillis());
+        clientConfig.property(ClientProperties.CONNECT_TIMEOUT, (int) config.connectTimeout().getMillis());
+        clientConfig.property(ClientProperties.READ_TIMEOUT, (int) config.socketTimeout().getMillis());
         return clientConfig;
     }
